feat: remove init plugin + add roadiz composables
