+++ composables/use-common-content.ts
@@ -3,20 +3,24 @@ import type { ValueOf } from '~/utils/types'
 
 export const COMMON_CONTENT_KEY = 'commonContent'
 
-export function useCommonContent() {
-    const commonContent = useNuxtData<CommonContent>(COMMON_CONTENT_KEY).data
-    const homeItem = computed(() => commonContent.value?.home)
-    const head = computed(() => commonContent.value?.head)
-    const errorWalker = computed(() => commonContent.value?.errorPage)
+export async function useCommonContent() {
+    const { data } = await useAsyncData<CommonContent>(COMMON_CONTENT_KEY, () => {
+        const fetch = useRoadizFetchFactory()
+
+        return fetch('/common_content')
+    })
+    const homeItem = computed(() => data.value?.home)
+    const head = computed(() => data.value?.head)
+    const errorWalker = computed(() => data.value?.errorPage)
 
     function getMenu(key: CommonContentMenuKey): ValueOf<CommonContent['menus']> | undefined {
-        return commonContent.value?.menus?.[key]
+        return data.value?.menus?.[key]
     }
 
     const mainMenuWalker = computed(() => getMenu('mainMenuWalker'))
 
     return {
-        commonContent,
+        data,
         head,
         homeItem,
         mainMenuWalker,
